Update api dependencies
This makes the following changes:

1. Remove dependencies in 'test.in' that are also listed in 'base.in'

These are not necessary to list again since 'test.in' already requires
the 'base.in' file.

---

2. Remove 'itsdangerous' from top-level dependencies.

This isn't a top-level dep and was likely pinned to solve an error back
in the day. Now modern versions of other packages require a more recent
version of this, and removing the constraint seem to fix it.

---

3. Constrain flask to < 2.0.

Flask greater than 2 has a conflict with the latest 'connexion' package.
The conflict is fixed on 'connexion' master, but there hasn't been a
release yet. So until that time we need to restrict flask to resolve the
conflict.

---

4. Use '--allow-unsafe' in 'pip-compile-multi'

This flag is poorly named and will actually become the default in
'pip-tools' soon. Using it can avoid some footguns around packages like
setuptools and pip.